task 2

+++ change.java
@@ -2,5 +2,5 @@ package com.example.task.com.example.task;
 
 @FunctionalInterface
 public interface change {
-    float changeRate(int m, float r);
+    float changeRate(float m, float r);
 }
